% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{best_results}
\alias{best_results}
\title{Get best parameter values and precision for dataset}
\usage{
best_results(dataset, predict_param, predict_param_type = "numeric")
}
\arguments{
\item{dataset}{data frame or numeric matrix. Experiment target.}

\item{predict_param}{numeric. Index of parameter to predict values of.}

\item{predict_param_type}{"numeric" or "character". If predict_param values are strings, then choose "character" to save it in predict_set
as such.}
}
\value{
best precision (percentage) and optimal k, metric and weighting scheme.
}
\description{
Empirically finds best parameter values and KNN precision for given dataset.
WARNING: this function may take A LOT of time to finish, especially for large datasets.
}
