% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{cross_validate}
\alias{cross_validate}
\title{Test KNN algorithm using cross-validation}
\usage{
cross_validate(dataset, parts, predict_param,
  predict_param_type = "numeric", k = "all", metric = "hassanat",
  weighting_scheme = "inverted")
}
\arguments{
\item{dataset}{data frame or numeric matrix. Cross-validation target.}

\item{parts}{numeric. Number of parts to divide data into}

\item{predict_param}{numeric. Index of parameter to predict values of.}

\item{predict_param_type}{"numeric" or "character". If predict_param values are strings, then choose "character" to save it in predict_set
as such.}

\item{k}{number of nearest neighbours to use for prediction or character string "all". Default: square root of number of
observations (rows) of training set.}

\item{metric}{character string giving a method for computing distances between vectors. One of:
"hassanat" (default), "canberra", "chebyshev", "chi_squared", "cosine", "euclidean", "euclidean_normalized", "euclidean_standarized",
"hamming", "mahalanobis", "manhattan", "pearson", "spearman".}

\item{weighting_scheme}{character string giving a method for weighting k nearest neighbours. One of:
"inverted" (default), "constant", "inverted_logarythmic".}
}
\value{
percentage of properly predicted values.
}
\description{
Divides provided into parts and performs cross-validation with given parameters.
}
